+++ src/runtime-wrapper/src/browser/websocket/client.ts
@@ -31,10 +31,18 @@ import type {
     WebSocketClientState,
     WebSocketConnectionMetrics
 } from "./types.js";
-import { DEFAULT_READINESS_POLL_INTERVAL_MS } from "./websocket-constants.js";
 
 const DEFAULT_MAX_QUEUE_SIZE = 100;
 const DEFAULT_FLUSH_INTERVAL_MS = 50;
+/**
+ * Default interval (milliseconds) between readiness polls when waiting for the
+ * GameMaker runtime to become ready before flushing pending patches.
+ *
+ * Lower values reduce the delay between runtime readiness and patch application,
+ * at the cost of slightly higher CPU usage during the polling window. The default
+ * of 50 ms is a reasonable balance for most hot-reload scenarios.
+ */
+export const DEFAULT_READINESS_POLL_INTERVAL_MS = 50;
 /**
  * Sentinel no-op used as the initial value for `detachWebSocketListeners` so
  * that disconnect/cleanup paths can always invoke it unconditionally. A real
